# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Veterinary Management
  description: Veterinary Management
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
    description: "Local server"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT authentication using OAuth2 Bearer token.

        **How to authenticate:**
        1. POST to `/loginJWT` with `administrator_id` and `password`
        2. Receive JWT token in response
        3. Include token in Authorization header: `Bearer <token>`

        **Available scopes:**
        - `RECEPTIONIST` - Clinic receptionist staff (manage visits, view schedules)
        - `CLINIC_MANAGER` - Clinic manager (full system access)
        - `VETERINARIAN` - Veterinarian (manage own schedule, view assigned visits)
        - `PET_OWNER` - Pet owner (view own pets and visits)
  parameters:
...
    visit_id:
      name: visit_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/VisitId'
    prescription_id:
      name: prescription_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/PrescriptionId'
    treatment_id:
      name: treatment_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/TreatmentId'
...
    pet_id:
      name: pet_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/PetId'
...
  requestBodies:
...
    VisitInput:
      description: "Visit input"
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              visit_date:
                $ref: '#/components/schemas/LocalDate'
              visit_time:
                $ref: '#/components/schemas/LocalTime'
              duration:
                type: integer
                description: "If omitted, defaults to 15 minutes"
              reason_for_visit:
                type: string
              price_per_fifteen:
                type: number
                format: double
              status:
                $ref: '#/components/schemas/VisitStatus'
              veterinarian_id:
                $ref: '#/components/schemas/VeterinarianId'
              pet_id:
                $ref: '#/components/schemas/PetId'
              pet_owner_id:
                type: integer
                format: int64
              diagnoses:
                type: string
              notes:
                type: string
    MedicationPrescriptionInput:
      description: "Medication prescription input"
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              quantity:
                type: integer
              dosage_instructions:
                type: string
              duration_in_days:
                type: integer
    TreatmentInput:
      description: "Treatment object input"
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              name:
                type: string
              description:
                type: string
              cost:
                type: number
                format: double
    VisitRescheduleInput:
      description: "Visit Reschedule input"
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              newDate:
                $ref: '#/components/schemas/LocalDate'
              newTime:
                $ref: '#/components/schemas/LocalTime'
              performedBy:
                type: string
    VisitDiagnosisInput:
      description: "Visit Diagnosis input"
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              diagnoses:
                type: string
              notes:
                type: string
...
  schemas:
    LocalTime:
      type: string
      format: time
      description: "Local time in the format HH:MM"
      default: "09:00"
    LocalDate:
      type: string
      format: date
      description: "Local date in the format YYYY-MM-DD"
      default: "2025-01-01"
    VeterinarianId:
      type: integer
      format: int64
      description: "Unique identifier for a Veterinarian"
...
    MedicationId:
      type: integer
      format: int64
      description: "Unique identifier for an Medication"
...
    VisitId:
      type: integer
      format: int64
      description: "Unique identifier for a Visit"
    PrescriptionId:
      type: integer
      format: int64
      description: "Unique identifier for a Medication Prescription"
    PetId:
      type: integer
      format: int64
      description: "Unique identifier for a Pet"
    PetOwnerId:
      type: integer
      format: int64
      description: "Unique identifier for a Pet Owner"
    TreatmentId:
      type: integer
      format: int64
      description: "Unique identifier for a Treatment"
...
    Visit:
      type: object
      properties:
        visit_id:
          $ref: '#/components/schemas/VisitId'
        visit_date:
          $ref: '#/components/schemas/LocalDate'
        visit_time:
          $ref: '#/components/schemas/LocalTime'
        duration:
          type: integer
        reason_for_visit:
          type: string
        diagnoses:
          type: string
        notes:
          type: string
        price_per_fifteen:
          type: double
        status:
          $ref: '#/components/schemas/VisitStatus'
        veterinarian_id:
          $ref: '#/components/schemas/VeterinarianId'
        pet_id:
          $ref: '#/components/schemas/PetId'
        pet_owner_id:
          $ref: '#/components/schemas/PetOwnerId'
        medication_prescriptions:
          type: array
          items:
            $ref: '#/components/schemas/MedicationPrescriptionInformation'
        treatment:
          $ref: '#/components/schemas/Treatment'
...
    MedicationPrescriptionInformation:
      type: object
      properties:
        prescription_id:
          $ref: '#/components/schemas/PrescriptionId'
        visit_id:
          $ref: '#/components/schemas/VisitId'
        medication_id:
          $ref: '#/components/schemas/MedicationId'
        medication_name:
          type: string
        quantity:
          type: integer
        dosage_instructions:
          type: string
        duration_in_days:
          type: integer
    Treatment:
      type: object
      properties:
        treatment_id:
          $ref: '#/components/schemas/TreatmentId'
        name:
          type: string
        description:
          type: string
        cost:
          type: number
          format: double
    TreatmentInformation:
      type: object
      properties:
        treatment_id:
          $ref: '#/components/schemas/TreatmentId'
        name:
          type: string
        description:
          type: string
        cost:
          type: number
          format: double
    VisitStatus:
      type: string
      enum: [ SCHEDULED, IN_PROGRESS, COMPLETED, CANCELLED, NOT_SHOWED_UP ]
      description: "Status of the visit"
...
    VisitHistoryEntry:
      type: object
      properties:
        consultation_date:
          $ref: '#/components/schemas/LocalDate'
        visit:
          $ref: '#/components/schemas/Visit'
        treatment:
          $ref: '#/components/schemas/TreatmentInformation'
        prescriptions:
          type: array
          items:
            $ref: '#/components/schemas/MedicationPrescriptionInformation'
paths:
...
  /visits:
    post:
      tags: [ Visits ]
      summary: "Create a visit"
      security:
        - bearerAuth: [RECEPTIONIST, PET_OWNER]
      requestBody:
        $ref: '#/components/requestBodies/VisitInput'
      responses:
        '201':
          description: "Visit created"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '400':
          description: "Invalid input"
        '404':
          description: "Veterinarian or Pet not found"
        '409':
          description: "Visit slot unavailable"
  /visits/walk-in:
    post:
      tags: [ Visits ]
      summary: "Create a walk-in visit"
      security:
        - bearerAuth: [ RECEPTIONIST ]
      parameters:
        - name: petId
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/PetId'
        - name: petOwnerId
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/PetOwnerId'
      responses:
        '201':
          description: "Visit created"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '404':
          description: "Pet or Pet Owner not found"
  /visits/{visit_id}:
    get:
      tags: [ Visits ]
      summary: "Get a Visit"
      security:
        - bearerAuth: [RECEPTIONIST, VETERINARIAN]
      parameters:
        - $ref: '#/components/parameters/visit_id'
      responses:
        '200':
          description: "Visit found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '404':
          description: "Visit not found"
    put:
      tags: [ Visits ]
      summary: "Update a Visit"
      security:
        - bearerAuth: [ RECEPTIONIST, VETERINARIAN ]
      parameters:
        - $ref: '#/components/parameters/visit_id'
      requestBody:
        $ref: '#/components/requestBodies/VisitInput'
      responses:
        '200':
          description: "Visit updated"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '400':
          description: "Invalid input"
        '404':
          description: "Visit not found"
        '409':
          description: "Visit slot unavailable"
    delete:
      tags: [ Visits ]
      summary: "Delete a Visit"
      security:
        - bearerAuth: [ RECEPTIONIST, VETERINARIAN ]
      parameters:
        - $ref: '#/components/parameters/visit_id'
      responses:
        '204':
          description: "Visit deleted"
        '404':
          description: "Visit not found"
  /visits/{visit_id}/start:
    post:
      tags: [ Visits ]
      summary: "Start a Visit consultation"
      security:
        - bearerAuth: [ VETERINARIAN ]
      parameters:
        - $ref: '#/components/parameters/visit_id'
      responses:
        '200':
          description: "Visit started"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '404':
          description: "Visit not found"
  /visits/{visit_id}/complete:
    post:
      tags: [ Visits ]
      summary: "Complete a Visit consultation"
      security:
        - bearerAuth: [ VETERINARIAN ]
      parameters:
        - $ref: '#/components/parameters/visit_id'
      responses:
        '200':
          description: "Visit started"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '404':
          description: "Visit not found"
  /visits/{visit_id}/diagnosis:
    put:
      tags: [ Visits ]
      summary: "Record diagnosis and notes for a visit"
      security:
        - bearerAuth: [ VETERINARIAN ]
      parameters:
        - $ref: '#/components/parameters/visit_id'
      requestBody:
        $ref: '#/components/requestBodies/VisitDiagnosisInput'
      responses:
        '204':
          description: "Diagnosis and notes recorded"
        '404':
          description: "Visit not found"
        '409':
          description: "Invalid visit status"
  /visits/{visit_id}/cancel:
    put:
      tags: [ Visits ]
      summary: "Cancel a visit"
      security:
        - bearerAuth: [ RECEPTIONIST, PET_OWNER ]
      parameters:
        - $ref: '#/components/parameters/visit_id'
      responses:
        '204':
          description: "Visit canceled"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '404':
          description: "Visit, Veterinarian or owner not found"
  /visits/{visit_id}/reschedule:
    put:
      tags: [ Visits ]
      summary: "Reschedule a visit"
      security:
        - bearerAuth: [ RECEPTIONIST, PET_OWNER ]
      parameters:
        - $ref: '#/components/parameters/visit_id'
      requestBody:
        $ref: '#/components/requestBodies/VisitRescheduleInput'
      responses:
        '204':
          description: "Visit rescheduled"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '404':
          description: "Visit or Veterinarian not found"
  /visits/{visit_id}/owner-not-showed-up:
    post:
      tags: [ Visits ]
      summary: "Mark a visit as owner not showed up"
      security:
        - bearerAuth: [ RECEPTIONIST ]
      parameters:
        - $ref: '#/components/parameters/visit_id'
      responses:
        '200':
          description: "Visit marked as not showed up"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '404':
          description: "Visit not found"
        '409':
          description: "Invalid visit status"

  # ============================================
  # Visit Prescriptions
  # ============================================
  /visits/{visit_id}/prescriptions:
    post:
      tags: [ Visits ]
      summary: "Create a medication prescription for a visit"
      security:
        - bearerAuth: [ VETERINARIAN ]
      parameters:
        - $ref: '#/components/parameters/visit_id'
        - name: medication_id
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/MedicationId'
      requestBody:
        $ref: '#/components/requestBodies/MedicationPrescriptionInput'
      responses:
        '201':
          description: "Medication prescription created"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicationPrescriptionInformation'
        '400':
          description: "Invalid input"
        '404':
          description: "Visit or Medication not found"
        '409':
          description: "Invalid visit status or insufficient stock"
    get:
      tags: [ Visits ]
      summary: "List medication prescriptions for a visit"
      security:
        - bearerAuth: [ VETERINARIAN ]
      parameters:
        - $ref: '#/components/parameters/visit_id'
      responses:
        '200':
          description: "Medication prescriptions found"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MedicationPrescriptionInformation'
        '404':
          description: "Visit not found"
  /visits/{visit_id}/prescriptions/{prescription_id}:
    get:
      tags: [ Visits ]
      summary: "Get a medication prescription"
      security:
        - bearerAuth: [ VETERINARIAN ]
      parameters:
        - $ref: '#/components/parameters/visit_id'
        - $ref: '#/components/parameters/prescription_id'
      responses:
        '200':
          description: "Medication prescription found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicationPrescriptionInformation'
        '404':
          description: "Visit or Medication prescription not found"
    put:
      tags: [ Visits ]
      summary: "Update a medication prescription"
      security:
        - bearerAuth: [ VETERINARIAN ]
      parameters:
        - $ref: '#/components/parameters/visit_id'
        - $ref: '#/components/parameters/prescription_id'
      requestBody:
        $ref: '#/components/requestBodies/MedicationPrescriptionInput'
      responses:
        '204':
          description: "Medication prescription updated"
        '400':
          description: "Invalid input"
        '404':
          description: "Visit or Medication prescription not found"
    delete:
      tags: [ Visits ]
      summary: "Delete a medication prescription"
      security:
        - bearerAuth: [ VETERINARIAN ]
      parameters:
        - $ref: '#/components/parameters/visit_id'
        - $ref: '#/components/parameters/prescription_id'
      responses:
        '204':
          description: "Medication prescription deleted"
        '404':
          description: "Visit or Medication prescription not found"
  /visits/prescriptions/report:
    get:
      tags: [ Visits ]
      summary: "List medications ordered by number of prescriptions in a period"
      security:
        - bearerAuth: [ ADMINISTRATOR ]
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
            description: "Start date of the period (ISO format: YYYY-MM-DD)"
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
            description: "End date of the period (ISO format: YYYY-MM-DD)"
      responses:
        200':
          description: "Medication prescription counts found"
          content:
            application/json:
              schema:
               type: array
               items:
                 type: object
               properties:
                 medicationId:
                   type: integer
                   format: int64
                 medicationName:
                   type: string
                 prescriptionCount:
                   type: integer
                   format: int64
        '400':
          description: "Invalid date range"
        '404':
          description: "No prescriptions found in the given period"
  /visits/veterinarians/report:
    get:
      tags: [ Visits ]
      summary: "List veterinarians ordered by number of prescriptions in a period"
      security:
        - bearerAuth: [ ADMINISTRATOR ]
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
            description: "Start date of the period (ISO format: YYYY-MM-DD)"
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
            description: "End date of the period (ISO format: YYYY-MM-DD)"
      responses:
        200':
          description: "veterinarians counts found"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                properties:
                  medicationId:
                    type: integer
                    format: int64
                  medicationName:
                    type: string
                  prescriptionCount:
                    type: integer
                    format: int64
        '400':
          description: "Invalid date range"
        '404':
          description: "No prescriptions found in the given period"
  # ============================================
  # Visit Treatments
  # ============================================
  /visits/treatments:
    post:
      tags: [ Treatments ]
      summary: "Create a new treatment"
      security:
        - bearerAuth: [ CLINIC_MANAGER ]
      requestBody:
        $ref: '#/components/requestBodies/TreatmentInput'
      responses:
        '201':
          description: "Treatment created"
        '400':
          description: "Invalid input"
    get:
      tags: [ Treatments ]
      summary: "List all treatments"
      security:
        - bearerAuth: [ RECEPTIONIST ]
      responses:
        '200':
          description: "List of treatments"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TreatmentInformation'
  /visits/treatments/{treatment_id}:
    get:
      tags: [ Treatments ]
      summary: "Get a treatment by ID"
      security:
        - bearerAuth: [ RECEPTIONIST ]
      parameters:
        - $ref: '#/components/parameters/treatment_id'
      responses:
        '200':
          description: "Treatment found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreatmentInformation'
        '404':
          description: "Treatment not found"
    put:
      tags: [ Treatments ]
      summary: "Update a treatment"
      security:
        - bearerAuth: [ CLINIC_MANAGER ]
      parameters:
        - $ref: '#/components/parameters/treatment_id'
      requestBody:
        $ref: '#/components/requestBodies/TreatmentInput'
      responses:
        '200':
          description: "Treatment updated"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreatmentInformation'
        '404':
          description: "Treatment not found"
    delete:
      tags: [ Treatments ]
      summary: "Delete a treatment"
      security:
        - bearerAuth: [ CLINIC_MANAGER ]
      parameters:
        - $ref: '#/components/parameters/treatment_id'
      responses:
        '204':
          description: "Treatment deleted"
        '404':
          description: "Treatment not found"
  /visits/{visit_id}/treatment:
    post:
      tags: [ Visits ]
      summary: "Assign a treatment to a visit"
      security:
        - bearerAuth: [ VETERINARIAN ]
      parameters:
        - $ref: '#/components/parameters/visit_id'
        - name: treatment_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '201':
          description: "Treatment assigned"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '404':
          description: "Visit or Treatment not found"
    get:
      tags: [ Visits ]
      summary: "Get the treatment assigned to a visit"
      security:
        - bearerAuth: [ RECEPTIONIST, VETERINARIAN ]
      parameters:
        - $ref: '#/components/parameters/visit_id'
      responses:
        '200':
          description: "Treatment found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreatmentInformation'
        '404':
          description: "Visit or Treatment not found"
    delete:
      tags: [ Visits ]
      summary: "Unassign a treatment from a visit"
      security:
        - bearerAuth: [ VETERINARIAN ]
      parameters:
        - $ref: '#/components/parameters/visit_id'
      responses:
        '200':
          description: "Treatment unassigned"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '404':
          description: "Visit not found"
...
  /visits/pets/{pet_id}/medical-history:
    get:
      tags: [ Visits ]
      security:
        - bearerAuth: [ VETERINARIAN, RECEPTIONIST ]
      parameters:
        - $ref: '#/components/parameters/pet_id'
      responses:
        '200':
          description: "Medical history retrieved"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VisitHistoryEntry'
        '404':
          description: "Pet not found"
        '400':
          description: "Invalid input"